refactor: move queue logic to utils
